refactor(api): move StructureIdentifier type in ValueObjects
