--- packages/api/src/@types/StructureIdentifiers.ts
@@ -1,4 +1,0 @@
-import AssociationIdentifier from "../valueObjects/AssociationIdentifier";
-import EstablishmentIdentifier from "../valueObjects/EstablishmentIdentifier";
-
-export type StructureIdentifier = AssociationIdentifier | EstablishmentIdentifier;
