+++ packages/api/src/@types/index.ts
@@ -1,5 +1,4 @@
 export * from "./Falsy";
-export * from "./StructureIdentifiers";
 export * from "./ParserPath";
 export * from "./ParserInfo";
 export * from "./ParserScheme";
