+++ packages/api/src/modules/etablissements/@types/EtablissementProvider.ts
@@ -1,6 +1,6 @@
 import { Etablissement } from "dto";
 import Provider from "../../providers/@types/IProvider";
-import { StructureIdentifier } from "../../../@types";
+import { StructureIdentifier } from "../../../valueObjects/@types/StructureIdentifier";
 
 export default interface EtablissementProvider extends Provider {
     isEtablissementProvider: boolean;
