+++ packages/api/src/modules/grant/grantExtract.service.ts
@@ -1,13 +1,13 @@
 import { SimplifiedEtablissement } from "dto";
 import csvStringifier from "csv-stringify/sync";
 import { BadRequestError } from "core";
-import { StructureIdentifier } from "../../@types";
 import associationsService from "../associations/associations.service";
 import AssociationIdentifier from "../../valueObjects/AssociationIdentifier";
 import EstablishmentIdentifier from "../../valueObjects/EstablishmentIdentifier";
 import GrantAdapter from "./grant.adapter";
 import { ExtractHeaderLabel } from "./@types/GrantToExtract";
 import grantService from "./grant.service";
+import { StructureIdentifier } from "../../valueObjects/@types/StructureIdentifier";
 
 class GrantExtractService {
     async buildCsv(identifier: StructureIdentifier): Promise<{ csv: string; fileName: string }> {
