+++ packages/api/src/modules/paymentFlat/paymentFlat.service.ts
@@ -3,12 +3,12 @@ import paymentFlatPort from "../../dataProviders/db/paymentFlat/paymentFlat.port
 import PaymentProvider from "../payments/@types/PaymentProvider";
 import AssociationIdentifier from "../../valueObjects/AssociationIdentifier";
 import EstablishmentIdentifier from "../../valueObjects/EstablishmentIdentifier";
-import { StructureIdentifier } from "../../@types";
 import { RawGrant, RawPayment } from "../grant/@types/rawGrant";
 import { ProviderEnum } from "../../@enums/ProviderEnum";
 import ProviderCore from "../providers/ProviderCore";
 import PaymentFlatEntity from "../../entities/PaymentFlatEntity";
 import PaymentFlatAdapter from "./paymentFlatAdapter";
+import { StructureIdentifier } from "../../valueObjects/@types/StructureIdentifier";
 
 export class PaymentFlatService extends ProviderCore implements PaymentProvider<PaymentFlatEntity> {
     constructor() {
