+++ packages/api/src/modules/providers/dauphin/dauphin.service.ts
@@ -12,7 +12,6 @@ import { asyncForEach } from "../../../shared/helpers/ArrayHelper";
 import DocumentProvider from "../../documents/@types/DocumentsProvider";
 import { RawApplication, RawGrant } from "../../grant/@types/rawGrant";
 import ProviderCore from "../ProviderCore";
-import { StructureIdentifier } from "../../../@types";
 import EstablishmentIdentifier from "../../../valueObjects/EstablishmentIdentifier";
 import AssociationIdentifier from "../../../valueObjects/AssociationIdentifier";
 import Siren from "../../../valueObjects/Siren";
@@ -21,6 +20,7 @@ import dauphinGisproPort from "../../../dataProviders/db/providers/dauphin/dauph
 import DauphinGisproDbo from "../../../dataProviders/db/providers/dauphin/DauphinGisproDbo";
 import DauphinSubventionDto from "./dto/DauphinSubventionDto";
 import DauphinDtoAdapter from "./adapters/DauphinDtoAdapter";
+import { StructureIdentifier } from "../../../valueObjects/@types/StructureIdentifier";
 
 export class DauphinService
     extends ProviderCore
