+++ packages/api/src/modules/providers/fonjep/fonjep.service.old.ts
@@ -8,7 +8,6 @@ import { FullGrantProvider } from "../../grant/@types/FullGrantProvider";
 import { FullGrantData, RawApplication, RawFullGrant, RawGrant, RawPayment } from "../../grant/@types/rawGrant";
 import ProviderCore from "../ProviderCore";
 import dataBretagneService from "../dataBretagne/dataBretagne.service";
-import { StructureIdentifier } from "../../../@types";
 import EstablishmentIdentifier from "../../../valueObjects/EstablishmentIdentifier";
 import AssociationIdentifier from "../../../valueObjects/AssociationIdentifier";
 import Siret from "../../../valueObjects/Siret";
@@ -18,6 +17,7 @@ import fonjepJoiner from "../../../dataProviders/db/providers/fonjep/fonjep.join
 import FonjepEntityAdapter from "./adapters/FonjepEntityAdapter.old";
 import FonjepSubventionEntity from "./entities/FonjepSubventionEntity.old";
 import FonjepPaymentEntity from "./entities/FonjepPaymentEntity.old";
+import { StructureIdentifier } from "../../../valueObjects/@types/StructureIdentifier";
 
 export enum FONJEP_SERVICE_ERRORS {
     INVALID_ENTITY = 1,
