+++ packages/api/src/modules/providers/scdl/scdl.grant.service.ts
@@ -3,7 +3,6 @@ import * as Sentry from "@sentry/node";
 import { AnyRawGrant, RawApplication, RawGrant } from "../../grant/@types/rawGrant";
 import { ProviderEnum } from "../../../@enums/ProviderEnum";
 import DemandesSubventionsProvider from "../../subventions/@types/DemandesSubventionsProvider";
-import { StructureIdentifier } from "../../../@types";
 import EstablishmentIdentifier from "../../../valueObjects/EstablishmentIdentifier";
 import AssociationIdentifier from "../../../valueObjects/AssociationIdentifier";
 import GrantProvider from "../../grant/@types/GrantProvider";
@@ -11,6 +10,7 @@ import miscScdlJoiner from "../../../dataProviders/db/providers/scdl/miscScdl.jo
 import MiscScdlGrantEntity from "./entities/MiscScdlGrantEntity";
 import MiscScdlGrantProducerEntity from "./entities/MiscScdlGrantProducerEntity";
 import MiscScdlAdapter from "./adapters/MiscScdl.adapter";
+import { StructureIdentifier } from "../../../valueObjects/@types/StructureIdentifier";
 
 export class ScdlGrantService implements DemandesSubventionsProvider<MiscScdlGrantProducerEntity>, GrantProvider {
     isGrantProvider = true;
