+++ packages/api/src/modules/subventions/@types/DemandesSubventionsProvider.ts
@@ -1,7 +1,7 @@
 import { DemandeSubvention } from "dto";
 import { RawApplication } from "../../grant/@types/rawGrant";
 import Provider from "../../providers/@types/IProvider";
-import { StructureIdentifier } from "../../../@types";
+import { StructureIdentifier } from "../../../valueObjects/@types/StructureIdentifier";
 
 export default interface DemandesSubvenentionsProvider<T> extends Provider {
     isDemandesSubventionsProvider: boolean;
