+++ packages/api/src/modules/subventions/subventions.service.ts
@@ -1,8 +1,8 @@
 import { DemandeSubvention } from "dto";
-import { StructureIdentifier } from "../../@types";
 import Flux from "../../shared/Flux";
 import { demandesSubventionsProviders } from "../providers";
 import { SubventionsFlux } from "./@types/SubventionsFlux";
+import { StructureIdentifier } from "../../valueObjects/@types/StructureIdentifier";
 
 export class SubventionsService {
     getDemandes(id: StructureIdentifier): Flux<SubventionsFlux> {
