+++ packages/api/src/valueObjects/@types/StructureIdentifier.ts
@@ -1,0 +1,4 @@
+import AssociationIdentifier from "../AssociationIdentifier";
+import EstablishmentIdentifier from "../EstablishmentIdentifier";
+
+export type StructureIdentifier = AssociationIdentifier | EstablishmentIdentifier;
